Clean up ghw
 - move the disk struct into types
 - move ghw mocker into its own subpackage so it can be reused
